TemplateURLPrepopulateData::GetEngineType moved.
This function now lives in search_engine_utils.h/cc which has its own
build source_set and is being used in blink.

Had to reshuffle our code so that Brave search engines can be accessed
in search_engine_utils.cc: moved our engines map into
brave_preopulated_engines.cc and added a function to get it.

Added a patch to add brave_prepopulated_engines.h/cc into
//components/search_engines:prepopulated_engines as additional sources.
To do that patched json_to_struct GN template.

Chromium change:

https://chromium.googlesource.com/chromium/src.git/+/2fd5742a51183821b2c7fbe139b962c40e8b99c8

commit 2fd5742a51183821b2c7fbe139b962c40e8b99c8
Author: Gayane Petrosyan <gayane@chromium.org>
Date:   Fri Oct 2 17:13:43 2020 +0000

    [SH-Blink] Adding histograms for Shared Highlights.

    Adding histograms to text fragment selector generation logic. Also
    adding a histogram that indicates if the origin of a link with text
    fragment is a known search engine.

    For finding out if a given url is a known search engine,
    I am separating template_url_prepopulate_data::GetEngineType and deps
    into a separate file to use it from blink/renderer/platform.
    Otherwise, template_url_prepopulate_data pulls deps that are also
    pulled by deps which somehow triggers
    AddressSanitizer: odr-violation(one definition rule) on some targets.
    See https://ci.chromium.org/p/chromium/builders/try/linux-libfuzzer-asan-rel/569669

    Unit tests for histograms in in a followup.
