+++ chromium_src/components/search_engines/template_url_prepopulate_data.cc
@@ -3,26 +3,25 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+#include "components/search_engines/template_url_prepopulate_data.h"
+
+#include <map>
+#include <vector>
+
 #include "base/stl_util.h"
 #include "base/strings/utf_string_conversions.h"
+#include "brave/components/search_engines/brave_prepopulated_engines.h"
 #include "components/country_codes/country_codes.h"
 
-// Pull in definitions for Brave prepopulated engines. It's ugly but these need
-// to be built as part of the search_engines static library.
-#include "../../../components/search_engines/brave_prepopulated_engines.cc"  // NOLINT
-#include "../../../components/search_engines/brave_prepopulated_engines.h"
-
 #define GetDataVersion GetDataVersion_ChromiumImpl
-#define GetEngineType GetEngineType_ChromiumImpl
 #if defined(OS_ANDROID)
 #define GetLocalPrepopulatedEngines GetLocalPrepopulatedEngines_Unused
 #endif
 #define GetPrepopulatedDefaultSearch GetPrepopulatedDefaultSearch_Unused
 #define GetPrepopulatedEngine GetPrepopulatedEngine_Unused
 #define GetPrepopulatedEngines GetPrepopulatedEngines_Unused
-#include "../../../../components/search_engines/template_url_prepopulate_data.cc"  // NOLINT
+#include "../../../../components/search_engines/template_url_prepopulate_data.cc"
 #undef GetDataVersion
-#undef GetEngineType
 #if defined(OS_ANDROID)
 #undef GetLocalPrepopulatedEngines
 #endif
@@ -36,51 +35,6 @@ void LocalizeEngineList(
     int country_id, std::vector<BravePrepopulatedEngineID>* engines);
 
 namespace {
-
-// Maps BravePrepopulatedEngineID to Chromium's PrepopulatedEngine.
-const std::map<BravePrepopulatedEngineID, const PrepopulatedEngine*>
-    brave_engines_map = {
-        {PREPOPULATED_ENGINE_ID_GOOGLE, &google},
-        {PREPOPULATED_ENGINE_ID_YAHOO, &brave_yahoo},
-        {PREPOPULATED_ENGINE_ID_YAHOO_AR, &brave_yahoo_ar},
-        {PREPOPULATED_ENGINE_ID_YAHOO_AT, &brave_yahoo_at},
-        {PREPOPULATED_ENGINE_ID_YAHOO_AU, &brave_yahoo_au},
-        {PREPOPULATED_ENGINE_ID_YAHOO_BR, &brave_yahoo_br},
-        {PREPOPULATED_ENGINE_ID_YAHOO_CA, &brave_yahoo_ca},
-        {PREPOPULATED_ENGINE_ID_YAHOO_CH, &brave_yahoo_ch},
-        {PREPOPULATED_ENGINE_ID_YAHOO_CL, &brave_yahoo_cl},
-        {PREPOPULATED_ENGINE_ID_YAHOO_CO, &brave_yahoo_co},
-        {PREPOPULATED_ENGINE_ID_YAHOO_DE, &brave_yahoo_de},
-        {PREPOPULATED_ENGINE_ID_YAHOO_DK, &brave_yahoo_dk},
-        {PREPOPULATED_ENGINE_ID_YAHOO_ES, &brave_yahoo_es},
-        {PREPOPULATED_ENGINE_ID_YAHOO_FI, &brave_yahoo_fi},
-        {PREPOPULATED_ENGINE_ID_YAHOO_FR, &brave_yahoo_fr},
-        {PREPOPULATED_ENGINE_ID_YAHOO_HK, &brave_yahoo_hk},
-        {PREPOPULATED_ENGINE_ID_YAHOO_ID, &brave_yahoo_id},
-        {PREPOPULATED_ENGINE_ID_YAHOO_IE, &brave_yahoo_ie},
-        {PREPOPULATED_ENGINE_ID_YAHOO_IN, &brave_yahoo_in},
-        {PREPOPULATED_ENGINE_ID_YAHOO_IT, &brave_yahoo_it},
-        {PREPOPULATED_ENGINE_ID_YAHOO_MX, &brave_yahoo_mx},
-        {PREPOPULATED_ENGINE_ID_YAHOO_MY, &brave_yahoo_my},
-        {PREPOPULATED_ENGINE_ID_YAHOO_NL, &brave_yahoo_nl},
-        {PREPOPULATED_ENGINE_ID_YAHOO_NO, &brave_yahoo_no},
-        {PREPOPULATED_ENGINE_ID_YAHOO_NZ, &brave_yahoo_nz},
-        {PREPOPULATED_ENGINE_ID_YAHOO_PE, &brave_yahoo_pe},
-        {PREPOPULATED_ENGINE_ID_YAHOO_PH, &brave_yahoo_ph},
-        {PREPOPULATED_ENGINE_ID_YAHOO_SE, &brave_yahoo_se},
-        {PREPOPULATED_ENGINE_ID_YAHOO_SG, &brave_yahoo_sg},
-        {PREPOPULATED_ENGINE_ID_YAHOO_TH, &brave_yahoo_th},
-        {PREPOPULATED_ENGINE_ID_YAHOO_TW, &brave_yahoo_tw},
-        {PREPOPULATED_ENGINE_ID_YAHOO_UK, &brave_yahoo_uk},
-        {PREPOPULATED_ENGINE_ID_YAHOO_VE, &brave_yahoo_ve},
-        {PREPOPULATED_ENGINE_ID_YAHOO_VN, &brave_yahoo_vn},
-        {PREPOPULATED_ENGINE_ID_BING, &bing},
-        {PREPOPULATED_ENGINE_ID_DUCKDUCKGO, &duckduckgo},
-        {PREPOPULATED_ENGINE_ID_DUCKDUCKGO_DE, &duckduckgo_de},
-        {PREPOPULATED_ENGINE_ID_DUCKDUCKGO_AU_NZ_IE, &duckduckgo_au_nz_ie},
-        {PREPOPULATED_ENGINE_ID_QWANT, &qwant},
-        {PREPOPULATED_ENGINE_ID_STARTPAGE, &startpage},
-};
 
 // Default order in which engines will appear in the UI.
 const std::vector<BravePrepopulatedEngineID> brave_engines_default = {
@@ -294,6 +248,8 @@ std::vector<const PrepopulatedEngine*> GetEnginesFromEngineIDs(
     BravePrepopulatedEngineID default_engine_id,
     size_t* default_search_provider_index = nullptr) {
   std::vector<const PrepopulatedEngine*> engines;
+  const auto& brave_engines_map =
+      TemplateURLPrepopulateData::GetBraveEnginesMap();
   for (size_t i = 0; i < engine_ids.size(); ++i) {
     const PrepopulatedEngine* engine = brave_engines_map.at(engine_ids.at(i));
     DCHECK(engine);
@@ -372,8 +328,10 @@ GetBravePrepopulatedEnginesForCountryID(
 // The intention of this function is to find the generic one
 // (ex: PREPOPULATED_ENGINE_ID_YAHOO) and then substitute the
 // country specific version.
-void LocalizeEngineList(
-    int country_id, std::vector<BravePrepopulatedEngineID>* engines) {
+// This function is not in the anonymous namespace because it
+// is used in brave_template_url_service_util_unittest.
+void LocalizeEngineList(int country_id,
+                        std::vector<BravePrepopulatedEngineID>* engines) {
   for (size_t i = 0; i < engines->size(); ++i) {
     if ((*engines)[i] == PREPOPULATED_ENGINE_ID_YAHOO) {
       const auto& it = yahoo_engines_by_country_id_map.find(country_id);
@@ -433,24 +391,6 @@ std::vector<std::unique_ptr<TemplateURLData>> GetLocalPrepopulatedEngines(
 }
 
 #endif
-
-SearchEngineType GetEngineType(const GURL& url) {
-  SearchEngineType type = GetEngineType_ChromiumImpl(url);
-  if (type == SEARCH_ENGINE_OTHER) {
-    for (const auto& entry : brave_engines_map) {
-      const auto* engine = entry.second;
-      if (SameDomain(url, GURL(engine->search_url))) {
-        return engine->type;
-      }
-      for (size_t j = 0; j < engine->alternate_urls_size; ++j) {
-        if (SameDomain(url, GURL(engine->alternate_urls[j]))) {
-          return engine->type;
-        }
-      }
-    }
-  }
-  return type;
-}
 
 // Functions below are copied verbatim from
 // components\search_engines\template_url_prepopulate_data.cc because they
