+++ components/search_engines/sources.gni
@@ -1,0 +1,9 @@
+# Copyright (c) 2020 The Brave Authors. All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at http://mozilla.org/MPL/2.0/.
+
+brave_components_search_engines_sources = [
+  "//brave/components/search_engines/brave_prepopulated_engines.cc",
+  "//brave/components/search_engines/brave_prepopulated_engines.h"
+]
