+++ patches/components-search_engines-BUILD.gn.patch
@@ -1,0 +1,12 @@
+diff --git a/components/search_engines/BUILD.gn b/components/search_engines/BUILD.gn
+index eb66a2d945650a81c3a3248975fb9deda036e51e..f11f019049297b7dd777a27b3a5bd4cd6a692784 100644
+--- a/components/search_engines/BUILD.gn
++++ b/components/search_engines/BUILD.gn
+@@ -194,6 +194,7 @@ json_to_struct("prepopulated_engines") {
+   schema_file = "prepopulated_engines_schema.json"
+   namespace = "TemplateURLPrepopulateData"
+   deps = [ ":search_engine_type" ]
++  additional_sources = brave_components_search_engines_sources
+ }
+ 
+ if (is_android) {
